Redesigned the user profile activity. Added delete functionality of items but  it works for a hardcoded user-id (Aviad's). Action bar transitions work for both my activities, and written for the others (but don't work since they don't exist yet). Changed ClothingItem API. All in all: now we can add items and delete them from a constant user.

+++ app/src/main/java/com/benefit/ui/profile/ClothingItem.java
@@ -1,13 +1,13 @@
 package com.benefit.ui.profile;
 
 /**
- * Represents a ui component of an item: the rounded rectangle with a photo and title.
+ * Represents a UI component of an item: the rounded rectangle with a photo and title.
  * Used in the RecyclerView as the object of the CardView.
  */
 public class ClothingItem {
 
 
-    private long mProductId;
+    private int mProductId;
     private String mImageUrl;
     private String mTitle;
 
@@ -15,7 +15,7 @@ public class ClothingItem {
     }
 
 
-    public ClothingItem(String imageUrl, String title, long id) {
+    public ClothingItem(String imageUrl, String title, int id) {
         mImageUrl = imageUrl;
         mTitle = title;
         mProductId = id;
@@ -37,7 +37,7 @@ public class ClothingItem {
         this.mTitle = mTitle;
     }
 
-    public long getmProductId() {
+    public int getmProductId() {
         return mProductId;
     }
